Allow selecting Hindered rotor fit in cantherm
Some time the best fit gives a negative energy barrier.
Therefore I added the ability to change the fit type to cosine
or fourier.

Also added an exception and error message for when this happens

Finally added more messages to the verbose log of cantherm to
help debug which rotor this occurs on.