Cope with duplicate species in an imported chemkin file.
At least the Dooley mechanism, and probably many more, have got the same
species under two different names. This commit allows you to import such
models. The 'original' chemkin file (and corresponding RMG dictionary) may
need some cleaning up before you can use the comparison tool, but I'll leave
that for now. The thermo.py file contains both, with a comment on one to say
that it's different; I'm not sure what happens when loading such a file.

(and moves the example to the top of the known species file for quicker debugging)