Do not add reverse reaction if it already exists.
If forward reaction has the same template, then do not include it; if not, then
we have found an additional transition state and the reaction should be included.
Do not include any reverse reactions if the family has its own reverse otherwise
we are including kinetics for more than one direction.