Revamp the way ScalarQuantities store and process uncertainties.
In commit 851b2ba42b60488e38267a0b8f231e6ee34cceff it was changed so that
Quantitiy.value and Quantity.uncertainty were now in customizable units,
and Quantitiy.value_si was the SI value. However, there was no uncertainty_si.

Now the uncertainty is also stored in SI units, like the value, and converted
to output units on demand. This change should be transparent, i.e. not change 
anything that was correct before, but it does create a new Quantity.uncertainty_si 
property, which is what we should have been using instead of .uncertainty in 
various places. i.e. we now need to go around using this to fix other bugs,
such as @GreenGroup/RMG-database#21 https://github.com/GreenGroup/RMG-database/issues/21

Also to do: do the same thing for the ArrayQuantity class.