Added documentation about kinetic database
All the information may not be as correct or as up to date as we'd like,
but at least it gives a base-line for new users to get used to the new
database. I'm not confident the specifics of the averaging
algorithm are correct. I'll also want to revisit the pitfalls sections for
further suggestions and checking if there is atom-by-atom children
checking.

Unfortunately, I'm leaving for a few months, so I wanted to just
upload whatever I had.