Partially fix transportDataTest unit test (and add another that fails)
Fixed it so the testJoback test works, and then added a species that
makes it fail. It was because of missing R=C=R ring group, but is now
just because we don't know what the correct values should be!
The test could be rewritten to make it just "don't crash on this species"