added Molecule.toSingleBonds() method
This returns a copy of the current molecule with only single bonds.
This is helpful for isomorphism comparisons with things that were made
via fromXYZ or via connectTheDots, which do not attempt to perceieve bond orders.
i.e. for use in checking the QM calculation made the right molecule.

Also created unit test for it.