Allow Energy units of 'Kelvins' for activation energies etc.
I wish there were a way to avoid this, but the way the chemkin parser is set up,
this looks like the best solution. I blame chemkin.

By allowing energies (like 831.4 kJ/mol) to be defined with declarations like (100.0,'K')
are we allowing unacceptable risks in terms of other bugs?
I hope not.
You must re-cythonize the quantities package after this commit.