Speed up the SMILES generation by looking up small molecules, and test it.
The previous algorithm checked the molecular formula anyway, so we may as well
use that to shortcut a bunch of things that can be uniquely identified that way,
to avoid calling OpenBabel or RDKit.

I also made a unit test that we get back the same SMILES we started with.

Inspired by and closes #175