Fix PDep Pruning bug introduced in e712697
Commit e712697f18c01e0d6ab419c37bc5248e1e3d3e8b changed the way we store PDep networks
so they are in a dictionary (keyed by entrance source) instead of a long list. This
greatly sped things up. However, it meant you have to iterate over two loops instead of one.
In modifying the code, some parts were not indented so were not included in the inner loop.
Hopefully this is not the only reason things sped up!! (although it must have helped!)

I noticed this while looking at the code around issue #176, which was just an off-by-one 
problem indexing an array from 1 instead of 0. The off-by-one thing doesn't actually hurt things
much because we don't use the array for anything anyway! (I'll remove it in the next commit)