Add message to UndeterminableKineticsError instead of printing warning msgs.
The warning msgs are misleading because on the first pass we try to find
a reaction template for the training reactions in the forward direction
as written.  However, these UndeterminableKineticsErrors are ignored
and the reactions stored in the reverse direction in case the user had
written them in the reverse direction.  By printing warnings at this
step we mislead the user into believing there is something wrong with
finding a template, when this is not actually the case.  The job will
fail if a reverse template cannot be found.