Added function isIdentical to groups.py
The function isIsomorphic was meant to compare molecules to groups,
therefore it respects wildcards (seeing as a molecule will never have an
atom named Cs or Ct). However, this makes the function isIsomorphic not
useful when tryint to compare two groups.

I've created a new function isIdentical which will not respect wildcards
and can tell if two groups are the same.