Try to fix a Conformer.getPrincipalMomentsOfInertia unit test.
It seems that numerical differences in the eigenvector search
can lead to the direction of the vector to switch, i.e.
we would be expecting -0.610114 in the second cell, but instead would
have +0.610114 in the 4th cell, and things would be swapped around.
Changing the starting coordinates by an insignificant amount
could lead to the other solution. It seems they're therefore
really equivalent?
I therefore changed the test so that either will pass.