Moved QM methods one level above the PM3 class.
We will specify the structure method elsewhere in the code,
but scriptAttempts and maxAttempts will be used by all
methods, whether pm3, pm6, or other methods we decide to
add in the future.

In gaussian.py this includes the inputFileKeywords method,
while we detail the keywords under a specific class.

In mopac.py this includes the keywords, while we define the
inputFileKeywords method under a specific class.