Add pm6 calculations to the QMthermo module.
This enables pm6 calculations in this module, however
there needs to be a discussion about how we want this implemented.

Currently the default is to use PM3, and this has not been
changed. Should another option be added to the qm settings
where the user specifies method? Should a user be able to
use any choice of structure method and basis set?

QM is set up to reuse outputs from previous calculations,
however with added methods, a check has to occur to ensure
that the data from a previous run used the same level of theory.