1,3-Hexadiene input file now saves QMfiles in the output directory.
This makes it slower to run, because it won't re-use old files if you
delete the output folder between runs, but that makes it a better test,
and this is one of our Makefile test targets (`make eg2`).
`make eg2` will delete the output directory before running, so we now wipe
the slate clean of old QMfiles.