Change the way cyclic and polycyclic corrections are handled AND don't include gauche correcitons.
Use overall ring correction instead of most destabilizing one for enthalpy.  (We were using
cyclopentane corrections for cyclopentene.)  Now it is corrected.

Don't include gauche corrections if the molecule is cyclic.  (Still need to include them
for parts of the molecule that aren't in the rings.. but we'll figure that out later.