Store the children of product groups when they are generated by RMG.
Product groups not found in the reaction family are automatically generated
by RMG for finding reverse reactions.  These groups don't have tracking
of parents and children like the rest of the groups.

Introduce this new functionality to track the children and parents
of the introduced groups.  i.e.

LogicOR Y
  - Group Y1
  - Group Y2
  - Group Y3

This way database testing can be done methodically.