Get it to work (mostly) with older versions of rdkit
Travis build would be faster if we could use the Ubuntu RPM
version of rdkit, but unfortunately that is old and doesn't know
about the element Dubnium. This hacky work-around solves it.
Helpful for people wanting to use Ubuntu, without QM.
(QM features require RDKit to be built with InChI support, which
means you have to build it from source anyway)