to/fromAdjacencyList: SYNTAX CHANGE: wildcard 'x' now includes 0.
I think it was probably meant to, but we should double check.
It looks to me like this wild card is not (yet) used anywhere
in the database anyway.

I also made it so we use the wildcard when writing adjacency lists
too, if either the unpaired electrons or pairs have [0,1,2,3,4]
they are written as 'x'.