Add test_kinetics_checkChildParentRelationship databaseTest
Implemented parent/child checking for nodes. I changed the way
matchNodeToChild in base.py worked:

-For the case where the parent is a LogicOr, it simply checks if the child
is in the components of the LogicOr

-There is also the interesting case where we have intermediate LogicOr
nodes e.g:

L3: GroupA
  L4: LogicOr
    L5: GroupB

In this case, we also need to check that GroupB is subgroup of GroupA. I
have implemented some extra lines for this in
test_kinetics_checkChildParentRelationship.

Also, we must ignore a check between the L4 LogicOr and L3 GroupA because
there is actually nothing to check. This was added to matchNodeToChild
to automatically pass this case.

I'm not convinced that we should have intermediate LogicOr groups like
this, but I've left the flexibility to do so for now.

Finally, I also changed the name of the function matchtoLogicOr to just
matchLogicOr as I thought it sounded more concise.