Remove outdated gprof2dot from external.
This package is now in the python package library and can be installed
via "pip install gprof2dot". It is in our requirements.txt file, so most
people will have that. That version is different, and our code now
reflects those changes, so this version is out of date and redundant.
Thus, I'm removing it.