Mark the StatMech group/tree test as Work In Progress
Per the discussion at https://github.com/GreenGroup/RMG-Py/pull/242
it's not clear if this test failure is really a problem or not,
or how best to resolve it. For now I'm marking as WIP so that 
we can get the rest of the tests in Master without Travis failing.