Correct gaussian environment variable when path is a list
This should not affect the environment variable that provides
a singe path to the g09 executable.

Some systems provide the environment variable as a list of
paths. To correct this, the environment variable
`GAUSS_EXEDIR`, which contains all these paths, is split
at the `:`. Each path is then searched for the g09 executable.