Consistent logging and source comments
For mopac and gaussian. Depending on the QM package used,
we were logging different lines and storing different source
formats. Not a good thing long term. This should help.

The sourcing will become important as we also load thermo files
created from previous calculations. If the job requests PM6 and the
previous file is a PM3 calculation, we currently use the old file.
Later commits will address this problem.