When loading Trees from database, do not sort them.
We now assume that the trees are in the best possible order, 
and this is maintained manually. Sometimes you want most specific
case first, in the case of overlapping nodes, but sometimes you 
want the smallest ring first (eg. if you have cyclic transition states)
which would mean the smallest functional group definition.
Enforcing a sorting algorithm seems risky to generalize, so we now
assume the tree was in the best order, and respect that.