Adding back spin orbit correction for thermo calculations
Adding back the SOC for all methods except G3. I also
exclude the methods for which @bbuesser added bond energy
correction, I am assuming he did not correct for the SOC
while fitting for those corrections.

Note: We are using CBS-Q corrections for CBS-QB3 method and
could be source of error. I checked the reference paper and
they fitted it to energies after adding SOC so we should do
the same.