Remove storing of kinetic comments from kinetic library upon loading.
Indication of which library the kinetics came from is already done
in chemkin when given a LibraryReaction.  The short and longDesc may
also be too long to be presented in chemkin files.  We don't use those
from the rules so we shouldn't use them here.  As long as the user
knows it's from a particular library, they can look up long commentary in
the database.