Fix calculateCyclicSymmetryNumber() function
Function was modifying original molecule's bond by changing them
to benzene bonds.  Now we make a copy of the molecule to start with (this uses
up a little extra memory but right now we just need to get it working.)

Also use a try and except block for converting to RDKit Mol, since it can
fail for some nitrogen cases.