Allow depository reactions to store reactants and products as Species.
This is easier for saving the dictionaries and looking at the reactions.. However
some of the copy functions don't work when the reaction has Species objects so we
need to change some of the code.  (This is probably something we need to fix in the
future so that reactions can be copied when their reactants and products are
Species rather than Molecule objects. But this is a faster fix.)