Avoid testing combinations of adjlists which produce malformed molecules.
@nickvandewiele 's isomorphismTest sometimes produces adjlists
that produce non-neutral molecules or molecules which give the incorrect
valence on certain atoms.  This was my quick hack for avoiding
tests which check these malformed molecules by passing an exception
on InvalidAjacencyListError and ValueError (when a non-neutral molecule
is detected.)  This is probably not the best method but is so far
an easy way to avoid errors.