Store wildcards as empty lists when retrieving lonePairs or charges from new-style adjlist.
This is to match the way wildcards are stored in fromOldAdjacencyList()
as well as the new behavior in the isomorphism checks.  It also saves
time when performing isomorphism checks where we don't need to iterate
through the list and check if the charge exists, we just just say it
matches the wildcard if it is an empty list.