Allow charges to be written in an adjlist for a group.  Don't write anything for wildcards.
Previously, toAdjacencyList() presumed that charges were not included
in groups (or that if they were zero, they were just not written down.)
Now we presume that charge is a wildcard if it is not written.  Otherwise
it has the same syntax as lonepairs, in other words if a single charge
is specified we can write

c+1

but if we have a list of charges we can have

c[-1,0,+1]

Note that we currently do not have built-in sorting for the charges
or lone pairs.  We also do not check if there are multiple same values yet.