Make CanTherm alert user if reacting species is not lowest E conformer
For large molecules, it's not always easy to know that the species you
use during a reaction is at it's lowest energy conformation. However,
the scan calculations that are used for inclusion of hindered rotors can
be used to alert a user if this is potentially the case, since most/all
users utilize the geometry from one of their reacting species to
initiate a scan calculation.

This commit now incorporated suggestions by @nickvandewiele
Merged and rebased by Richard West <r.west@neu.edu>