1-2)make cantherm alert user if reacting species is not lowest E conformer 3) small change to input doc for rmg families 4-5) included CanTherm example and some related unittests
make cantherm alert user if reacting species is not lowest E conformer

For large molecules, it's not always easy to know that the species you
use during a reaction is at it's lowest energy conformation. However,
the scan calculations that are used for inclusion of hindered rotors can
be used to alert a user if this is potentially the case, since most/all
users utilize the geometry from one of their reacting species to
initiate a scan calculation.

update to #287

incorporated suggestions by @nickvandewiele

small change to input doc for rmg familes

just made it a little clearer that the 'default' option is just a
string, while multiple Families are designated as a list of strings

included CanTherm example and some related unittests

The cantherm example is for those who want to make an input file from
scratch. Cantherm has few unittests, so i included some, as a start.
Also updated/changed comments/doc relating to loadChemkinFile as it was
unclear when i first read it.

undo undesired changes to two files

Perhaps someone can squash this commit with several previous ones...