Refactor the pruning algorithm a little bit.
This should be a bit faster, and a bit clearer.
Faster because we have fewer lookups of self.edge.species[index]
and because we can stop iterating as soon as we reach a small enough edge.

Now that this has been debugged, I'm not sure we need the logging statement
on line 1101, because the same info is logged about 15 lines further down...