Attempt to clarify #290 re. Symmetry corrections in HBI radical GA estimates.
I changed a method name, added an assertion, and added some text to the docstring.
Just trying to clarify what is currently being done, without changing anything.

My remaining concern is with how this interacts with QMThermo, where the "stable" 
(saturated parent) molecule is estimated via PM6 or something. This current 
implementation assumes that the the entropy has not been corrected for symmetry
(because that is what it now does for group additive thermo).