Small CanTherm bug fix with a couple updates to documenation
CanTherm was crashing for cases when users specified TS frequencies in
the input file. I also updated part of the documentation for input
files. I think the kinetics function example was for Measure, and if
used with RMG-Py, would cause cantherm to crash. Added more doc relevant
to new T-parameter functionality implemented by Shamel, and a
Troubleshooting/FAQ section with one entry.