Cantherm fix, doc update, more bondE corrections
Cantherm was crashing if the input file did not have a model chemistry
with a bond energy correction dictionary. Creation of an initial empty
dictionary fixes this. Also added B3LYP bond energy corrections  from
the same table and source as the CBS-QB3 corrections. Made some small
updates to the doc.