mimic the main() method instead of deprecated Main() class of gprof2dot
see
https://code.google.com/p/jrfonseca/source/detail?r=bf25a2abb09bf2cb821832d76900f0dd8326eb8a&repo=gprof2dot

for more info on this change of gprof2dot.

The main method is copied, except the optparser part that reads from
sys.arg. The output of optparser is replaced by an empty Options class
populated with hard-coded options.

Only gprof2dot imports to a limit number of classes, attributes were made.