use try..except statement to reduce verbosity of parametrized tests
the mechanism uses a try block in which a test method is executed.
Only in the case of an AssertionError, or a generic Exception, the
respective methods failed(*args) and exception(*args) will be called and
produce output. If the test succeeds, nothing will be printed.

Moreover, we don't need nose-parametrized, we can just call the callback
without passing it to a decorator!