Fix isMoleculeForbidden() check in database/base.py.
Only declare a molecule to be forbidden if all the labels from the forbidden
group exist in the molecule.  Previously, RMG was failing the following
molecule:

1     C u0 p0 c0 {2,D} {3,S} {6,S}
2     C u0 p0 c0 {1,D} {4,S} {7,S}
3     C u0 p0 c0 {1,S} {5,D} {8,S}
4  *1 O u0 p2 c0 {2,S} {9,S}
5     N u0 p0 c+1 {3,D} {10,D}
6     H u0 p0 c0 {1,S}
7     H u0 p0 c0 {2,S}
8     H u0 p0 c0 {3,S}
9  *2 H u0 p0 c0 {4,S}
10    N u0 p2 c-1 {5,D}

on this forbidden group in the Substitution_O family:

quadrad_singlet
1 *3 [C,N,Si] u0 p2

because the 10th atom nitrogen has 2 lone pairs and 0 unpaired electrons.
This alerted to the fact that mapping checks are always valid when initialMaps
are empty, which is not the behavior we want.  If any label in the forbidden group
cannot be found in the molecule, assume that the group is NOT forbidden and proceed
to the rest of the test.