Don't regenerate reactions from libraries since they're already added to the edge.
Found that RMG attempts to add library reactions to the edge through the
generateReactions function.  This is not needed because the reaction libraries are
already added to the edge in the RMG class's initialize step.  (It also caused problems
when we wanted to retain duplicate reactions in the libraries, we end up with too
many copies.)

This should reduce duplicate processes in the code.