Update qm/molecule with "new" methods from RMG
The getRadicalCount method is pretty old now, but it wasn't
updated here. Instead of going through the list of atoms, we
just call it to get the rad number.

Also, updated the default folders method. The previous code
would set the scratch and permanent folders to be the same.

Conflicts:
	rmgpy/qm/molecule.py