Begin setup of TS searches using nwchem
For now, the NEB search can be done using nwchem, but further
steps need to be finished. No IRC is required since NEB is a path
analysis method. The Zero-Temp String method also needs to
be implemented.

The direct method needs testing and more work. The pre-optimization
needs to be fixed as the freezing of atoms isn't working yet.
The IRC (MEPGS) is implemented and should run, but has not been tested.