Change PdepArrhenius and MultiPDepArrhenius classes to allow MultiArrhenius objects
Let MultiArrhenius objects replace Arrhenius ones in the list stored
in PdepArrhenius.arrhenius.  This allows the correct PLOG behavior for duplicate
kinetics at a single pressure.  This will then evaluate the duplicate
kinetics defined by MultiArrhenius at that pressure and use that rate for
klow or khigh.

All that is required is to change the class setting for Arrhenius to a more relaxed
parent definition: KineticsModel. Nothing else needs to be changed.