Add small comment as to why residual function returns delta, 1.
The residual function should return Delta, IRES to the DASPK function.
The following is copied from the DASPK documentation.

C         Depending on the value of INFO(20), RES may also be used
C         to define the residuals of the sensitivity equations.
C         If INFO(20) = 2 (sensitivity equations defined analytically)
C         and if IRES = 0, RES should evaluate only the Ny residuals of
C         the original DAE. If INFO(20) = 2 and IRES = 1, RES should
C         evaluate the residuals of of the sensitivity equations, storing
C         them in DELTA immediately following the residuals of the state
C         variables. If the parallel method is also selected (INFO(27)>1)
C         for sensitivity computation, RES should evaluate the residuals
C         of the sensitivity equations with respect to parameters starting
C         from INFO(26)+1, and the ones every other INFO(27) as
C         INFO(26)+1, INFO(26)+1+INFO(27),...,INFO(26)+1+j*INFO(27),...
C         where the values of INFO(26) and INFO(27) are generated by a
C         set-up routine of the message passing interface (MPI).
C
C         For the initialization of a Hessenberg index-2 DAE problem,
C         if INFO(20) = 2 and IRES = 2, RES is used to evaluate
C         the residuals of the time derivatives for index-2 constraints
C         which should include the index-2 sensitivity constraints if
C         sensitivity is considered (see the description of item INFO(11)
C         in the documentation).
C
C         Depending on the value of INFO(28), RES may also be used to
C         define the residuals of the equations for the quadrature variables.
C         If INFO(28)--the number of quadrature variables--is greater than 0,
C         the RES routine should evaluate only the residuals of the quadrature
C         equations if IRES = 3. The residuals for the quadrature equations
C         should be placed in the last part of the Ny components of the DELTA
C         array for the state variables. If INFO(20) = 2 is also chosen, the
C         corresponding sensitivity equations to the quadrature equations
C         should also be evaluated.
C
C         The INFO(28) should also include the number of output variables
C         (i.e., it should be the sum of the number of quadrature variables
C         and the number of output variables). The residual equations for
C         the output variables MUST be scaled by CJ, i.e. transform
C             delta = y_out - f(y)
C         into
C             delta = CJ*(y_o - f(y))
C         The initial values of the output variables are specified by
C         the user, not calculated inside DDASPK even if the initialization
C         is selected by the user. The quadrature and output variables can
C         be mixed together without any specific order.
C
C         Overall, an efficient RES routine has the following multiple-branch
C         selection statement:
C
C            IF (IRES .EQ. 0) THEN
C               define the residuals for the state equations
C               (excluding quadrature/output equations)
C            ELSE IF (IRES .EQ. 1) THEN
C               define the residuals for the sensitivity equations
C               (excluding quadrature/output sensitivity equations)
C            ELSE IF (IRES .EQ. 2) THEN
C               define the residuals of the time derivatives of the
C               index-2 constraints and/or index-2 sensitivity constraints
C            ELSE IF (IRES .EQ. 3) THEN
C               define the residuals for the quadrature equations and/or
C               the corresponding sensitivity equations
C            END IF