Fix to add labels to libraries in loadOld
Previously we listed the adjlist for all reactants and products in the
entry for the reaction. A recent change now has a separate dictionary file
and reaction files, so that reactants can be editted once instead of once
for every reaction they are in.

However, now we need to add a label property to each reaction entry to
substitute for the the adjList. This fix adds one line which adds the
correct label (hopefully in all cases)