Molecule().getAromaticResonanceIsomers now checks rings of any size and atom type.
Previously only 6-membered rings of carbon were checked.
Now, any sized ring, of any atom, is checked. If RDKit says it's an aromatic 
bond, then we make it an aromatic bond. If all the bonds in the ring are aromatic,
then we add it to the list of isomers returned.

We test that the oxygen in a furan ring is turned into an Ob AtomType.
This tests both the getAromaticResonanceIsomers and the updateAtomTypes at once,
which is not really the best unit test, but is better than nothing...
More (and better) unit tests should be added.