Molecule().isAromatic() now returns True for any sized ring of any Xb.
Previously only 6-membered rings of Cb or Cbf were allowed.
Now any sized ring, and any Xb or Xbf is allowed, as long as
everything in the ring is Xb or Xbf. (i.e. Ob, Cb, Cbf, N3b, etc.)

Perhaps we should also check Huckel rule of 4N+2 electrons in the ring?
I suppose a manually crafted (via inputting an adjacency list) 5-membered
ring of Cb atoms would pass this test, but be wrong.
However, such a molecule would presumably not be made by RDKit, so should
not show up via getAromaticResonanceIsomers(). So we're probably OK for now...

We test that furan is detected as aromatic.