Fixed lone pairs bug in Platts group additivity for solvation.
Modified procedures to change lone pairs to radicals based on the atom's
valency, and then saturate the radicals with hydrogens.

When unsaturating, we re-add a radical correction to the H-bonding
parameter 'A'. Eventually, a radical library will be added for this,
but for now, every Oxygen radical that had been saturated will be
corrected with a value of -0.345, which is the value for -OH and
-OOH.

Added a couple of unit tests.