Remove getKineticsFromRules().  Was a duplicate function and gave mixed kinetics comments.
When you find an exact match for a template, you get kinetics from getKineticsFromRules, but
nothing for estimates...
Inside getKinetics(), fetching the kinetics is done later through getKineticsForTemplate,
which does the same thing but for both exact matches AND estimates.  So we were getting
the correct answer but mixed kinetic comments in the chemkin. It's unecessary.