Enlarge the model with each species INDIVIDUALLY, rather than multiple at once
Because the reactionsMovedFromEdge in rmgpy.rmg.model.enlarge
is often incorrectly calculated since the numOldEdgeReactions was from the
initial number of Edge reactions, not after each species added iteration.
So this resulted in a bug of sometimes not printing all the edge species
properly.  As well as not accounting for edge duplicates correctly which we
now need.

I also think that doing it in this manner is more clean and easier to track the
effect of an individual species.  We do get a bunch of additional printouts
in the beginning from the inerts, but I feel that this is ok and adds transparency.
In the case where someone has an inert that still reacts in a reaction library
they'd want to see it individually added.