Eliminate duplicated R_Recombination hardcoding for picking kinetics in family.py
The reversing of templates is already done in the rmgpy.data.kinetics.rules
function getAllRules.  So having this hardcoding for R_Recombination outside
of that in rmgpy.data.kinetics.family is completely redundent and also poorly
organized.

Putting in getAllRules makes sense because it means that we consider the
reverse template to simply mean the same thing as a duplicate entry with alternate
kinetics and treats it as such.