Change the way HBI method is handled for radical thermo for both QM and GAVs.
Use a priority system which first prioritizes by method:
1) Library saturated value
2) QM saturated value
3) Group additivity value

It then sorts the radical's thermo first by the priority, then by lowest
H298 value.  This ensures that a GAV or QM value, even if lower, does not
override a more accurate library value.

It then sorts the resonance isomers in self.molecule by this priority.

Of course, if the radical itself was found in the library, it would
produce that value prior to making it to this step.