Add an ``strict`` argument to descendTree function and related functions to enforce atom label matching.
We use strict atom label matching in finding the correct kinetic groups to associate
with a reaction and its reaction template.  However, the previous method used for LogicOr groups
was flawed.  It found the very first child node of an LogicOR group and used its atomLabels
to match the reactant.  However, sometimes the LogicOR group contained children that had
different atomLabels.  This resulted in some correct matches being thrown out.

By adding a ``strict`` argument to the descendTree function, the matchNodeToStructure, and
matchToStructure functions, we can be sure that full atom label matching takes place.  Note
that this is currently used primarily for kinetic groups and is not implemented for other
functions such as thermo estimation.